pyproject.toml
[project]
name = "enpm667-project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=1.26,<2",
    "pandas>=2.3.3",
    "roboticstoolbox-python>=1.1.1",
    "scipy>=1.16.3",
    "spatialmath-python>=1.1.15",
    "tqdm>=4.67.1",
]
